python-api-backup-20250627-013319/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.30.0
openai-whisper
ollama>=0.5.0
pydub>=0.25.0
numpy>=1.24.0
python-multipart>=0.0.12
aiofiles>=24.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
httpx>=0.27.0
pydantic>=2.9.0
jinja2>=3.1.0